remove tests for original batch reducer logic
